Replaced a LexicalData object with a more narrower-scope InflectionData. Split class declarations into several files
