+++ index.js
@@ -1,4 +1,7 @@
-export * from './lib/lib'
-export * from './presenter/presenter'
-export { dataSet as LatinDataSet } from './lib/lang/latin/latin'
-export { dataSet as GreekDataSet } from './lib/lang/greek/greek'
+import InflectionData from './lib/inflection-data'
+import LanguageData from './lib/language-data'
+import { dataSet as LatinDataSet } from './lib/lang/latin/latin'
+import { dataSet as GreekDataSet } from './lib/lang/greek/greek'
+import Presenter from './presenter/presenter'
+
+export { InflectionData, LanguageData, LatinDataSet, GreekDataSet, Presenter }
