+++ lib/inflection-data.js
@@ -1,0 +1,42 @@
+import * as Models from 'alpheios-data-models'
+import Suffix from './suffix'
+import Footnote from './footnote'
+
+/**
+ * A return value for inflection queries
+ */
+export default class InflectionData {
+  constructor (language) {
+    // this.homonym = homonym
+    this.language = language
+    this[Models.Feature.types.part] = [] // What parts of speech are represented by this object.
+  }
+
+  static readObject (jsonObject) {
+    // let homonym = Models.Homonym.readObject(jsonObject.homonym)
+
+    let lexicalData = new InflectionData()
+    lexicalData[Models.Feature.types.part] = jsonObject[Models.Feature.types.part]
+
+    for (let part of lexicalData[Models.Feature.types.part]) {
+      let partData = jsonObject[part]
+      lexicalData[part] = {}
+
+      if (partData.suffixes) {
+        lexicalData[part].suffixes = []
+        for (let suffix of partData.suffixes) {
+          lexicalData[part].suffixes.push(Suffix.readObject(suffix))
+        }
+      }
+
+      if (partData.footnotes) {
+        lexicalData[part].footnotes = []
+        for (let footnote of partData.footnotes) {
+          lexicalData[part].footnotes.push(Footnote.readObject(footnote))
+        }
+      }
+    }
+
+    return lexicalData
+  }
+}
