+++ presenter/views/greek.js
@@ -1,22 +1,24 @@
-import * as Lib from '../../lib/lib'
+import languages from '../../lib/languages'
 import * as Greek from '../../lib/lang/greek/greek'
-import * as View from '../lib/view'
+import View from '../lib/view'
+import GroupFeatureType from '../lib/group-feature-type'
+import Table from '../lib/table'
 
-class GreekView extends View.View {
+class GreekView extends View {
   constructor () {
     super()
-    this.language = Lib.languages.greek
+    this.language = languages.greek
 
     /*
-    Default grammatical features of a view. It child views need to have different feature values, redefine
+    Default grammatical features of a View. It child views need to have different feature values, redefine
     those values in child objects.
      */
     this.features = {
-      numbers: new View.GroupFeatureType(Greek.numbers, 'Number'),
-      cases: new View.GroupFeatureType(Greek.cases, 'Case'),
-      declensions: new View.GroupFeatureType(Greek.declensions, 'Declension'),
-      genders: new View.GroupFeatureType(Greek.genders, 'Gender'),
-      types: new View.GroupFeatureType(Greek.types, 'Type')
+      numbers: new GroupFeatureType(Greek.numbers, 'Number'),
+      cases: new GroupFeatureType(Greek.cases, 'Case'),
+      declensions: new GroupFeatureType(Greek.declensions, 'Declension'),
+      genders: new GroupFeatureType(Greek.genders, 'Gender'),
+      types: new GroupFeatureType(Greek.types, 'Type')
     }
   }
 
@@ -25,7 +27,7 @@ class GreekView extends View.View {
    * an inflection table differently.
    */
   createTable () {
-    this.table = new View.Table([this.features.declensions, this.features.genders,
+    this.table = new Table([this.features.declensions, this.features.genders,
       this.features.types, this.features.numbers, this.features.cases])
     let features = this.table.features
     features.columns = [Greek.declensions, Greek.genders, Greek.types]
@@ -83,7 +85,7 @@ class NounViewSimplified extends NounView {
   }
 
   static suffixCellFilter (suffix) {
-    return suffix.extendedLangData[Lib.languages.greek].primary
+    return suffix.extendedLangData[languages.greek].primary
   }
 }
